separando classes conta
