+++ src/conta/conta.model.ts
@@ -1,7 +1,8 @@
 import { TipoConta } from 'src/enums/tipo-conta.enum';
 import { v4 as uuidv4 } from 'uuid';
+import { ContaInterface } from './conta.interface';
 
-export abstract class Conta {
+export abstract class Conta implements ContaInterface {
   public id: string;
   constructor(
     public saldo: number,
@@ -23,51 +24,4 @@ export abstract class Conta {
   abstract verificarSaldoInsuficiente(valor: number): void;
 
   abstract transferir(destino: Conta, valor: number): void;
-}
-
-export class ContaCorrente extends Conta {
-  private limiteChequeEspecial: number = 100;
-
-  constructor(saldo: number, clienteID: string) {
-    super(saldo, clienteID, TipoConta.CORRENTE);
-  }
-
-  verificarSaldoInsuficiente(valor: number): void {
-    if (valor > this.saldo + this.limiteChequeEspecial) {
-      throw new Error('Saldo insuficiente');
-    }
-  }
-
-  transferir(destino: Conta, valor: number): void {
-    let saldoTotal = this.saldo + this.limiteChequeEspecial;
-
-    this.verificarSaldoInsuficiente(valor);
-
-    saldoTotal -= valor;
-    destino.saldo += valor;
-  }
-}
-
-export class ContaPoupanca extends Conta {
-  private taxaJuros: number = 0.01;
-
-  constructor(saldo: number, clienteID: string) {
-    super(saldo, clienteID, TipoConta.POUPANCA);
-  }
-
-  transferir(destino: Conta, valor: number): void {
-    this.verificarSaldoInsuficiente(valor);
-    this.saldo -= valor;
-    destino.saldo += valor;
-  }
-
-  calcularTaxa(): number {
-    return this.saldo * this.taxaJuros;
-  }
-
-  verificarSaldoInsuficiente(valor: number): void {
-    if (valor > this.saldo) {
-      throw new Error('Saldo insuficiente');
-    }
-  }
 }
